+++ NEWS.md
@@ -4,6 +4,7 @@
 * Allow `tar_quarto_rep()` to write reports to subdirectories with the help of a project-level `_quarto.yml` (#211, @lgaborini).
 * `tar_map2*()` functions now aggregate dynamic branches in parallel over static branches (#213).
 * `tar_map2*()` functions gain an `unlist` argument.
+* Call `parallel::clusterExport()` in `make_psock_cluster()` to make sure globals carry over to parallel socket clusters.
 
 # tarchetypes 0.12.0